backend/requirements.txt
fastapi
sqlalchemy
pydantic
python-dotenv
google-cloud-speech
google-cloud-language
python-multipart
uvicorn
pyttsx3

